+++ R/theme.R
@@ -1,9 +1,24 @@
 
 #' Forest plot default theme
 #'
+#' @description
+#'
 #' Default theme for the forest plot, but can pass other parameters. The
 #' parameters will be passed to corresponding elements of the forest plot.
-#' See \code{\link[grid]{gpar}} for details.
+#'
+#' \itemize{
+#'   \item{\code{ci_*}}{Control the graphical parameters of confidence intervals}
+#'   \item{\code{legend_*}}{Control the graphical parameters of legend}
+#'   \item{\code{xaxis_*}}{Control the graphical parameters of x-axis}
+#'   \item{\code{refline_*}}{Control the graphical parameters of reference line}
+#'   \item{\code{vertline_*}}{Control the graphical parameters of vertical line}
+#'   \item{\code{summary_*}}{Control the graphical parameters of diamond shaped summary CI}
+#'   \item{\code{footnote_*}}{Control the graphical parameters of footnote}
+#'   \item{\code{title_*}}{Control the graphical parameters of title}
+#'   \item{\code{arrow_*}}{Control the graphical parameters of arrow}
+#' }
+#'
+#' See \code{\link[grid]{gpar}} for more details.
 #'
 #' @param base_size The size of text
 #' @param base_family The font family
@@ -42,12 +57,25 @@
 #' @param title_cex Multiplier applied to font size for title.
 #' @param title_fontface The font face for title, default is \code{'bold'}.
 #' @param title_col Color of title.
-#' @param title_fontfamily Font family of title. 
+#' @param title_fontfamily Font family of title.
+#' @param arrow_type Type of the arrow below x-axis, see \code{\link[grid]{arrow}}.
+#' @param arrow_label_just The justification of the arrow label relative to arrow. Control
+#' the arrow label to align to the starting point of the arrow \code{"start"} (default) or
+#' the ending point of the arrow \code{"end"}.
+#' @param arrow_length The length of the arrow head, default is \code{0.05}.
+#' See \code{\link[grid]{arrow}}.
+#' @param arrow_lwd Line width of the arrow, same as \code{xaxis_lwd} by default.
+#' @param arrow_fill Filling color of the arrow head, default is \code{"black"}.
+#' @param arrow_col Line and text color of the arrow, same as \code{arrow_fill} by default.
+#' @param arrow_cex Multiplier applied to font size for arrow label, same as \code{xaxis_cex} by default.
 #' @param ... Other parameters passed to table. See \code{\link[gridExtra]{tableGrob}}
 #'  for details.
 #'
-#' @importFrom utils modifyList
 #'
+#' @importFrom utils modifyList
+#' @seealso \code{\link[gridExtra]{tableGrob}} \code{\link{forest}} \code{\link[grid]{textGrob}}
+#'  \code{\link[grid]{gpar}} \code{\link[grid]{arrow}} \code{\link[grid]{segmentsGrob}}
+#' \code{\link[grid]{linesGrob}} \code{\link[grid]{pointsGrob}} \code{\link[grid]{legendGrob}}
 #' @return A list.
 #'
 #' @export
@@ -88,11 +116,20 @@ forest_theme <- function(base_size = 12,
                          title_fontface = "bold",
                          title_col = "black",
                          title_fontfamily = base_family,
-                         # Legend
+                         # Arrow
+                         arrow_type = c("open", "closed"),
+                         arrow_label_just = c("start", "end"),
+                         arrow_length = 0.05,
+                         arrow_lwd = xaxis_lwd,
+                         arrow_fill = "black",
+                         arrow_col = arrow_fill,
+                         arrow_cex = xaxis_cex,
                          # legend_lwd = 0.6,
                          ...){
 
     legend_position <- match.arg(legend_position, c("right", "top", "bottom", "none"))
+
+    if (!is.unit(arrow_length)) arrow_length <- unit(arrow_length, units = "inches")
 
     # Default color set
     col_set <- c("#e41a1c","#377eb8","#4daf4a","#984ea3","#ff7f00",
@@ -168,6 +205,17 @@ forest_theme <- function(base_size = 12,
                                col = title_col,
                                fontfamily = title_fontfamily))
 
+    # Arrow
+    arrow <- list(type = match.arg(arrow_type),
+                  label_just = match.arg(arrow_label_just),
+                  length = arrow_length,
+                  gp = gpar(fontsize = base_size,
+                            fontfamily = base_family,
+                            lwd = arrow_lwd,
+                            fill = arrow_fill,
+                            col = arrow_col,
+                            cex = arrow_cex))
+
     # Table body
     core <- list(fg_params = list(hjust = 0,
                                x = 0.05,
@@ -196,6 +244,7 @@ forest_theme <- function(base_size = 12,
                 xaxis = xaxis_gp,
                 footnote = footnote_gp,
                 title  = title_gp,
+                arrow = arrow,
                 refline = refline_gp,
                 vertline = vertline_gp,
                 summary = sum_gp,
@@ -204,7 +253,7 @@ forest_theme <- function(base_size = 12,
 }
 
 
-# 
+#
 make_group_theme <- function(theme, group_num){
 
   # Default color set
@@ -238,7 +287,7 @@ make_group_theme <- function(theme, group_num){
     # Check for group and legend label
     if(group_num > 1 & length(theme$legend$label) < group_num & length(theme$legend$label) > 1)
       stop("More groups than legend labels.")
-    
+
     return(theme)
 
 }